+++ spec/models/activity_spec.rb
@@ -41,6 +41,12 @@ RSpec.describe Activity, type: :model do
       it { should validate_presence_of(:sector) }
     end
 
+    context "when planned dates are blank" do
+      subject { build(:activity, planned_start_date: nil, planned_end_date: nil, wizard_status: :dates) }
+      it { should validate_presence_of(:planned_start_date) }
+      it { should validate_presence_of(:planned_end_date) }
+    end
+
     context "when status is blank" do
       subject { build(:activity, status: nil, wizard_status: :status) }
       it { should validate_presence_of(:status) }
@@ -48,12 +54,12 @@ RSpec.describe Activity, type: :model do
 
     context "when planned_start_date is blank" do
       subject { build(:activity, planned_start_date: nil, wizard_status: :dates) }
-      it { should_not validate_presence_of(:planned_start_date) }
+      it { should validate_presence_of(:planned_start_date) }
     end
 
     context "when planned_end_date is blank" do
       subject { build(:activity, planned_end_date: nil, wizard_status: :dates) }
-      it { should_not validate_presence_of(:planned_end_date) }
+      it { should validate_presence_of(:planned_end_date) }
     end
 
     context "when actual_start_date is blank" do
@@ -86,6 +92,40 @@ RSpec.describe Activity, type: :model do
       end
     end
 
+    context "when the actual_start_date is not blank" do
+      it "allows todays date" do
+        activity = build(:activity, actual_start_date: Date.today)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "allows dates in the past" do
+        activity = build(:activity, actual_start_date: 1.year.ago)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "does not allow a date in the future" do
+        activity = build(:activity, actual_start_date: 1.day.from_now)
+        expect(activity.valid?).to be_falsey
+      end
+    end
+
+    context "when the actual_end_date is not blank" do
+      it "allows todays date" do
+        activity = build(:activity, actual_end_date: Date.today)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "allows dates in the past" do
+        activity = build(:activity, actual_end_date: 1.year.ago)
+        expect(activity.valid?).to be_truthy
+      end
+
+      it "does not allow a date in the future" do
+        activity = build(:activity, actual_end_date: 1.day.from_now)
+        expect(activity.valid?).to be_falsey
+      end
+    end
+
     context "when recipient_region is blank" do
       subject { build(:activity, recipient_region: nil, wizard_status: :country) }
       it { should validate_presence_of(:recipient_region) }
@@ -112,8 +152,8 @@ RSpec.describe Activity, type: :model do
       it { should validate_presence_of(:description) }
       it { should validate_presence_of(:sector) }
       it { should validate_presence_of(:status) }
-      it { should_not validate_presence_of(:planned_start_date) }
-      it { should_not validate_presence_of(:planned_end_date) }
+      it { should validate_presence_of(:planned_start_date) }
+      it { should validate_presence_of(:planned_end_date) }
       it { should_not validate_presence_of(:actual_start_date) }
       it { should_not validate_presence_of(:actual_end_date) }
       it { should validate_presence_of(:recipient_region) }