+++ CHANGELOG.md
@@ -53,3 +53,5 @@
 - Users can report an activities implementing organisation
 - Users can edit a reported implementing organisation
 - Users can view implementing organisations in the Activity XML
+- Planned start and end dates are mandatory
+- Actual start and end dates must not be in the future